Feature testing and bug fixes (#194)
* Updated a failing execution test: exponentiation directory.

* Updated failing execution tests related to keccak256, sha256, ripemd160.
Used the tests to verify the relevant issue (#args, padding).

* Bug fix: in type conversion of constants, due to unbound integer type.

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/builtinFunctions/keccak256_multiple_arguments.sol
@@ -1,9 +1,9 @@
 contract c {
-    function foo(uint a, uint b, uint c) public returns (bytes32 d) {
+    function foo(uint256 a, uint256 b, uint256 c) public returns (bytes32 d) {
         d = keccak256(abi.encodePacked(a, b, c));
     }
 }
 // ====
 // compileViaYul: also
 // ----
-// foo(uint,uint,uint): 0xa, 0xc, 0xd -> 0x00bc740a98aae5923e8f04c9aa798c9ee82f69e319997699f2782c40828db9fd81
+// foo(uint256,uint256,uint256): 0x0a, 0x0c, 0x0d -> 0x00bc740a98aae5923e8f04c9aa798c9ee82f69e319997699f2782c40828db9fd81
